% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_connect.R
\name{get_metrics_db_conn}
\alias{get_metrics_db_conn}
\title{Connect to Metrics Database}
\usage{
get_metrics_db_conn(env_file = ".env")
}
\arguments{
\item{env_file}{path to environment file containing keys}
}
\description{
Connect to Metrics Database
}
