% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricsReport-package.R
\docType{package}
\name{metricsReport-package}
\alias{metricsReport}
\alias{metricsReport-package}
\title{metricsReport: R package used for reporting metrics for RCDS}
\description{
The package provides functions used for retrieving, cleaning, and reporting metrics for the Research Computing and Data Services (RCDS) team.
}
\keyword{internal}
