% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workshop_specific.R
\name{set_topic}
\alias{set_topic}
\title{Set Topic}
\usage{
set_topic(d, topic_categories, category_order = NULL, blank_category = "")
}
\arguments{
\item{d}{workshop dataset}

\item{topic_categories}{a character vector whose name is the column and the value is the corresponding topic}
}
\description{
Set Topic
}
