% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting_helpers.R
\name{make_df_time_plot}
\alias{make_df_time_plot}
\title{Plot a Column's Breakdown Over Time}
\usage{
make_df_time_plot(input_df, year_array, col)
}
\arguments{
\item{input_df}{a data frame containing at least the columns `fis_year_`,
`person_id`, and `col`}

\item{year_array}{a vector of fiscal years to include as columns}

\item{col}{the name (as a string) of the categorical column to break down}
}
\value{
a ggplot object
}
\description{
Builds on [make_df_time_table()] and reshapes the result into a line plot
of percentage over fiscal year, one line per level of `col`.
}
